# apps/admin-api/src/support/api-docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin-permission-presets": {
      "post": {
        "description": "여러 권한을 묶은 프리셋을 생성합니다. 프리셋이 적용된 관리자에게는 즉시 반영됩니다 (참조형).\n\n**Required Permissions (ALL):** `PERMISSION_PRESET_MANAGE` (권한 프리셋 생성/수정/삭제/조회)",
        "operationId": "AdminPermissionPresetController_createPreset",
        "parameters": [
          {
            "required": true,
            "name": "x-user-lang",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "en-US",
                "ko",
                "fr",
                "es",
                "de",
                "ja",
                "ms"
              ]
            },
            "description": "사용자의 언어 코드"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreatePresetReq"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "SUCCESS"
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminCreatePresetRes"
                    },
                    "error": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "result",
                    "data",
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "이미 사용중인 프리셋 코드",
            "content": {
              "application/json": {
                "examples": {
                  "example1": {
                    "summary": "이미 사용중인 프리셋 코드",
                    "value": {
                      "result": "ERROR",
                      "data": null,
                      "error": {
                        "code": "PERMISSION_PRESET_CODE_DUPLICATE",
                        "message": "이미 사용중인 프리셋 코드",
                        "data": null
                      }
                    }
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "ERROR"
                      ]
                    },
                    "data": {
                      "type": "null",
                      "nullable": true
                    },
                    "error": {
                      "$ref": "#/components/schemas/AdminApiErrorMessage"
                    }
                  },
                  "required": [
                    "result",
                    "data",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Access JWT": []
          }
        ],
        "summary": "권한 프리셋 생성",
        "tags": [
          "Admin Permission Preset"
        ]
      },
      "get": {
        "description": "등록된 모든 프리셋을 조회합니다. 각 프리셋의 권한 개수가 함께 반환됩니다.\n\n**Required Permissions (ALL):** `PERMISSION_PRESET_MANAGE` (권한 프리셋 생성/수정/삭제/조회)",
        "operationId": "AdminPermissionPresetController_getPresetList",
        "parameters": [
          {
            "required": true,
            "name": "x-user-lang",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "en-US",
                "ko",
                "fr",
                "es",
                "de",
                "ja",
                "ms"
              ]
            },
            "description": "사용자의 언어 코드"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "SUCCESS"
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminGetPresetListRes"
                    },
                    "error": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "result",
                    "data",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Access JWT": []
          }
        ],
        "summary": "권한 프리셋 목록 조회",
        "tags": [
          "Admin Permission Preset"
        ]
      }
    },
    "/v1/admin-permission-presets/{id}": {
      "get": {
        "description": "프리셋 정보와 포함된 권한 코드 목록을 조회합니다.\n\n**Required Permissions (ALL):** `PERMISSION_PRESET_MANAGE` (권한 프리셋 생성/수정/삭제/조회)",
        "operationId": "AdminPermissionPresetController_getPreset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "프리셋 ID",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "required": true,
            "name": "x-user-lang",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "en-US",
                "ko",
                "fr",
                "es",
                "de",
                "ja",
                "ms"
              ]
            },
            "description": "사용자의 언어 코드"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "SUCCESS"
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminGetPresetRes"
                    },
                    "error": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "result",
                    "data",
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않는 권한 프리셋",
            "content": {
              "application/json": {
                "examples": {
                  "example1": {
                    "summary": "존재하지 않는 권한 프리셋",
                    "value": {
                      "result": "ERROR",
                      "data": null,
                      "error": {
                        "code": "PERMISSION_PRESET_NOT_FOUND",
                        "message": "존재하지 않는 권한 프리셋",
                        "data": null
                      }
                    }
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "ERROR"
                      ]
                    },
                    "data": {
                      "type": "null",
                      "nullable": true
                    },
                    "error": {
                      "$ref": "#/components/schemas/AdminApiErrorMessage"
                    }
                  },
                  "required": [
                    "result",
                    "data",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Access JWT": []
          }
        ],
        "summary": "권한 프리셋 단건 조회",
        "tags": [
          "Admin Permission Preset"
        ]
      },
      "patch": {
        "description": "프리셋의 이름/설명/권한 구성을 수정합니다.\n권한 코드 목록은 전체 교체 방식으로 동작합니다 (요청에 포함된 코드만 남고 나머지는 제거).\n이 프리셋이 적용된 모든 관리자에게는 즉시 반영됩니다 (참조형).\n\n**Required Permissions (ALL):** `PERMISSION_PRESET_MANAGE` (권한 프리셋 생성/수정/삭제/조회)",
        "operationId": "AdminPermissionPresetController_updatePreset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "프리셋 ID",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "required": true,
            "name": "x-user-lang",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "en-US",
                "ko",
                "fr",
                "es",
                "de",
                "ja",
                "ms"
              ]
            },
            "description": "사용자의 언어 코드"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdatePresetReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "SUCCESS"
                      ]
                    },
                    "data": {
                      "type": "null"
                    },
                    "error": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "result",
                    "data",
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않는 권한 프리셋",
            "content": {
              "application/json": {
                "examples": {
                  "example1": {
                    "summary": "존재하지 않는 권한 프리셋",
                    "value": {
                      "result": "ERROR",
                      "data": null,
                      "error": {
                        "code": "PERMISSION_PRESET_NOT_FOUND",
                        "message": "존재하지 않는 권한 프리셋",
                        "data": null
                      }
                    }
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "ERROR"
                      ]
                    },
                    "data": {
                      "type": "null",
                      "nullable": true
                    },
                    "error": {
                      "$ref": "#/components/schemas/AdminApiErrorMessage"
                    }
                  },
                  "required": [
                    "result",
                    "data",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Access JWT": []
          }
        ],
        "summary": "권한 프리셋 수정",
        "tags": [
          "Admin Permission Preset"
        ]
      },
      "delete": {
        "description": "프리셋을 삭제합니다 (soft delete). 이 프리셋이 적용되어 있던 관리자들은 해당 프리셋이 자동으로 적용 해제됩니다.\n\n**Required Permissions (ALL):** `PERMISSION_PRESET_MANAGE` (권한 프리셋 생성/수정/삭제/조회)",
        "operationId": "AdminPermissionPresetController_deletePreset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "프리셋 ID",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "required": true,
            "name": "x-user-lang",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "en-US",
                "ko",
                "fr",
                "es",
                "de",
                "ja",
                "ms"
              ]
            },
            "description": "사용자의 언어 코드"
          }
        ],
        "responses": {
          "200": {
            "description": "삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "SUCCESS"
                      ]
                    },
                    "data": {
                      "type": "null"
                    },
                    "error": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "result",
                    "data",
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않는 권한 프리셋",
            "content": {
              "application/json": {
                "examples": {
                  "example1": {
                    "summary": "존재하지 않는 권한 프리셋",
                    "value": {
                      "result": "ERROR",
                      "data": null,
                      "error": {
                        "code": "PERMISSION_PRESET_NOT_FOUND",
                        "message": "존재하지 않는 권한 프리셋",
                        "data": null
                      }
                    }
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "ERROR"
                      ]
                    },
                    "data": {
                      "type": "null",
                      "nullable": true
                    },
                    "error": {
                      "$ref": "#/components/schemas/AdminApiErrorMessage"
                    }
                  },
                  "required": [
                    "result",
                    "data",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Access JWT": []
          }
        ],
        "summary": "권한 프리셋 삭제",
        "tags": [
          "Admin Permission Preset"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nx Nest Starter Admin API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Access JWT": {
        "scheme": "bearer",
        "name": "Access JWT",
        "type": "http",
        "in": "header",
        "description": "Access JWT for user authentication",
        "x-apidog": {
          "authType": "bearer",
          "addTokenTo": "header",
          "headerPrefix": "Bearer",
          "token": "{{bearerToken}}"
        }
      }
    },
    "schemas": {
      "AdminApiErrorMessage": {
        "type": "object",
        "properties": {
          "traceId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "traceId",
          "code",
          "message"
        ]
      },
...
      "AdminCreatePresetRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "생성된 프리셋 ID"
          }
        },
        "required": [
          "id"
        ]
      },
      "AdminCreatePresetReq": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "description": "프리셋 코드 (대문자/숫자/언더스코어, 대문자로 시작)",
            "example": "ROOT"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "프리셋 표시명",
            "example": "루트 관리자"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 500,
            "description": "프리셋 설명"
          },
          "permissionCodes": {
            "type": "array",
            "description": "프리셋에 포함할 권한 코드 목록",
            "example": [
              "USER_LIST",
              "USER_READ"
            ],
            "items": {
              "type": "string",
              "enum": [
                "ADMIN_ACCOUNT_MANAGE",
                "PERMISSION_PRESET_MANAGE",
                "ADMIN_PERMISSION_ASSIGN",
                "ADMIN_PERMISSION_LIST",
                "USER_LIST",
                "USER_READ",
                "USER_SUSPEND",
                "MAIL_SEND",
                "PUSH_SEND"
              ]
            }
          }
        },
        "required": [
          "code",
          "name",
          "permissionCodes"
        ]
      },
      "AdminPresetListItemRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "프리셋 ID"
          },
          "code": {
            "type": "string",
            "description": "프리셋 코드"
          },
          "name": {
            "type": "string",
            "description": "프리셋 이름"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "프리셋 설명"
          },
          "permissionCount": {
            "type": "integer",
            "description": "포함된 권한 수"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "생성 일시"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "description",
          "permissionCount",
          "createdAt"
        ]
      },
      "AdminGetPresetListRes": {
        "type": "object",
        "properties": {
          "list": {
            "description": "프리셋 목록",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminPresetListItemRes"
            }
          }
        },
        "required": [
          "list"
        ]
      },
      "AdminGetPresetRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "프리셋 ID"
          },
          "code": {
            "type": "string",
            "description": "프리셋 코드"
          },
          "name": {
            "type": "string",
            "description": "프리셋 이름"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "프리셋 설명"
          },
          "permissionCodes": {
            "description": "포함된 권한 코드 목록",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "생성 일시"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "description",
          "permissionCodes",
          "createdAt"
        ]
      },
      "AdminUpdatePresetReq": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "프리셋 표시명"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 500,
            "description": "프리셋 설명"
          },
          "permissionCodes": {
            "type": "array",
            "description": "프리셋에 포함할 권한 코드 목록 (전체 교체)",
            "example": [
              "USER_LIST",
              "USER_READ"
            ],
            "items": {
              "type": "string",
              "enum": [
                "ADMIN_ACCOUNT_MANAGE",
                "PERMISSION_PRESET_MANAGE",
                "ADMIN_PERMISSION_ASSIGN",
                "ADMIN_PERMISSION_LIST",
                "USER_LIST",
                "USER_READ",
                "USER_SUSPEND",
                "MAIL_SEND",
                "PUSH_SEND"
              ]
            }
          }
        },
        "required": [
          "name",
          "permissionCodes"
        ]
      },
...
    }
  }
}